Use homebrew Lst implementation for parameter lists

+++ compiler/src/dotty/tools/dotc/core/tasty/NameBuffer.scala
@@ -71,8 +71,8 @@ class NameBuffer extends TastyBuffer(10000) {
     val encodedValue = paramSig match {
       case paramSig: TypeName =>
         nameRefs(paramSig.toTermName).index
-      case paramSig: Int =>
-        -paramSig
+      case paramSig: Integer =>
+        -paramSig.intValue
     }
     writeInt(encodedValue)
   }
